+++ tests/testthat/test_AnnotateIDVCF.R
@@ -5,15 +5,15 @@ test_that("AnnotateIDVCF function with hg19", {
   stopifnot(requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5"))
   load("testdata/test_AnnotateIDVCF.Rdata")
   id.vcf <- ReadStrelkaIDVCF("testdata/Strelka-ID-GRCh37/Strelka.ID.GRCh37.vcf")
-  df <- 
+  list <- 
     AnnotateIDVCF(id.vcf, 
                   ref.genome = 
                     BSgenome.Hsapiens.1000genomes.hs37d5::BSgenome.Hsapiens.1000genomes.hs37d5)
-  df1 <- AnnotateIDVCF(id.vcf, ref.genome = "GRCh37")
-  df2 <- AnnotateIDVCF(id.vcf, ref.genome = "hg19")
-  expect_equal(df, strelka.ID.vcf.GRCh37)
-  expect_equal(df, df1)
-  expect_equal(df, df2)
+  list1 <- AnnotateIDVCF(id.vcf, ref.genome = "GRCh37")
+  list2 <- AnnotateIDVCF(id.vcf, ref.genome = "hg19")
+  expect_equal(list$annotated.vcf, strelka.ID.vcf.GRCh37)
+  expect_equal(list$annotated.vcf, list1$annotated.vcf)
+  expect_equal(list$annotated.vcf, list2$annotated.vcf)
 })
 
 test_that("AnnotateIDVCF with hg38", {
@@ -21,13 +21,13 @@ test_that("AnnotateIDVCF with hg38", {
   stopifnot(requireNamespace("BSgenome.Hsapiens.UCSC.hg38"))
   load("testdata/test_AnnotateIDVCF.Rdata")
   id.vcf <- ReadStrelkaIDVCF("testdata/Strelka.ID.GRCh38.vcf")
-  df3 <- AnnotateIDVCF(id.vcf, 
+  list3 <- AnnotateIDVCF(id.vcf, 
                        ref.genome = 
                          BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38)
-  df4 <- AnnotateIDVCF(id.vcf, ref.genome = "GRCh38")
-  df5 <- AnnotateIDVCF(id.vcf, ref.genome = "hg38")
-  expect_equal(df3, strelka.ID.vcf.GRCh38)
-  expect_equal(df3, df4)
-  expect_equal(df3, df5)
+  list4 <- AnnotateIDVCF(id.vcf, ref.genome = "GRCh38")
+  list5 <- AnnotateIDVCF(id.vcf, ref.genome = "hg38")
+  expect_equal(list3$annotated.vcf, strelka.ID.vcf.GRCh38)
+  expect_equal(list3$annotated.vcf, list4$annotated.vcf)
+  expect_equal(list3$annotated.vcf, list5$annotated.vcf)
 })